requirements.txt
flask
yfinance
feedparser
requests
notion-client
gunicorn
schedule
python-dotenv
pandas
google-generativeai
PyPDF2
beautifulsoup4
